% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_azure_storage.R
\name{download_provider_data}
\alias{download_provider_data}
\title{Download Provider Data}
\usage{
download_provider_data(
  file,
  data_path = file.path("app_data"),
  inputs_data_version = envvars$NHP_INPUTS_DATA_VERSION,
  ...
)
}
\arguments{
\item{file}{The name of the file to download.}

\item{data_path}{The path to the app's data directory.}

\item{inputs_data_version}{The version of the inputs data to use.}

\item{...}{Additional arguments to pass to `AzureStor::download_adls_file()`.}
}
\value{
The path to the downloaded file.
}
\description{
Download a selected type of provider data from ADLS and save it as a parquet
file in the app's data directory.
}
