% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_azure_storage.R
\name{get_adls_fs}
\alias{get_adls_fs}
\title{Get ADLS Filesystem}
\usage{
get_adls_fs()
}
\value{
the adls filesystem
}
\description{
Get the Azure Data Lake Storage filesystem using the endpoint and container
specified in environment variables. (AZ_STORAGE_EP and AZ_STORAGE_CONTAINER)
}
