% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_inequalities_utils.R
\name{inequality_choices_to_display}
\alias{inequality_choices_to_display}
\title{Convert inequality choices to display case}
\usage{
inequality_choices_to_display(snake_choice)
}
\arguments{
\item{snake_choice}{Character vector of snake_case names e.g., 'zero_sum'}
}
\value{
Character vector of display names. Returns \code{NA} for
  unrecognized values.
}
\description{
Convert inequality choices to display case
}
\examples{
inequality_choices_to_display("zero_sum")

}
