% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_inequalities_utils.R
\name{inequality_choices_to_snake}
\alias{inequality_choices_to_snake}
\title{Convert inequality choices to snake case}
\usage{
inequality_choices_to_snake(display_choice)
}
\arguments{
\item{display_choice}{Character vector of display names e.g., Zero-sum"}
}
\value{
Character vector of snake_case names. Returns \code{NA} for
  unrecognized values.
}
\description{
Convert inequality choices to snake case
}
\examples{
inequality_choices_to_snake("Zero-sum")

}
