% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_inequalities_utils.R
\name{initialise_hrg_table}
\alias{initialise_hrg_table}
\title{Initialise inequality choice table}
\usage{
initialise_hrg_table(provider_inequalities, params)
}
\arguments{
\item{provider_inequalities}{A data frame containing provider inequality data.}

\item{params}{The standard params object}
}
\value{
A tibble with two columns:
  \item{hrg_code}{Character vector of unique HRG codes from the input data}
  \item{choice}{Character vector indicating the inequality adjustment choice
    for each HRG code.}
    Possible values are defined in get_inequality_choice_mappings()
}
\description{
Creates a tibble containing unique HRG codes for selected provider and
a "choice" column indicating the type of adjustment to apply.
Set to "No change" by default
}
\details{
If editing an existing scenario, previous parameters will be loaded.
}
