% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_azure_storage.R
\name{load_inequalities_data}
\alias{load_inequalities_data}
\title{Get inequalities data}
\usage{
load_inequalities_data(
  inputs_data_version = Sys.getenv("NHP_INPUTS_DATA_VERSION", "dev")
)
}
\arguments{
\item{inputs_data_version}{The version of the inputs data to use.}
}
\value{
A tibble.
}
\description{
Read the parquet file for inequalities
}
