% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_provider_data.R
\name{load_provider_data}
\alias{load_provider_data}
\title{Get Provider Data}
\usage{
load_provider_data(file, data_path = file.path("app_data"))
}
\arguments{
\item{file}{The name of the file to read.}

\item{data_path}{The path to the directory containing the parquet files.}
}
\value{
A tibble.
}
\description{
Read the parquet file containing a selected type of provider data.
}
