% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_azure_storage.R
\name{get_adls_fs}
\alias{get_adls_fs}
\title{Get ADLS Filesystem}
\usage{
get_adls_fs()
}
\value{
the adls filesystem
}
\description{
try to use a managed token, this will only work if run in an Azure data centre
if it fails, try instead to use Azure Resource Manager
}
