% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_azure_storage.R
\name{get_all_data_files}
\alias{get_all_data_files}
\title{Get All Data Files}
\usage{
get_all_data_files(
  inputs_data_version = Sys.getenv("NHP_INPUTS_DATA_VERSION", "dev")
)
}
\arguments{
\item{inputs_data_version}{The version of the inputs data to use.}
}
\description{
Download all provider data files from ADLS and save them as parquet files in
the app's data directory.
}
