% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_provider_data.R
\name{load_provider_data}
\alias{load_provider_data}
\title{Get Provider Data}
\usage{
load_provider_data(file, data_path = app_sys("app", "data"))
}
\arguments{
\item{file}{The name of the file to read.}

\item{inputs_data_version}{The version of the inputs data to use.}
}
\value{
A tibble.
}
\description{
Read the parquet file containing a selected type of provider data.
}
