% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_azure_storage.R
\name{load_provider_data}
\alias{load_provider_data}
\title{Get Provider Data}
\usage{
load_provider_data(
  file,
  inputs_data_version = Sys.getenv("NHP_INPUTS_DATA_VERSION", "dev")
)
}
\arguments{
\item{file}{The name of the file to read.}

\item{inputs_data_version}{The version of the inputs data to use.}
}
\value{
A tibble.
}
\description{
Read the parquet file containing a selected type of provider data.
}
